+++ clang/lib/AST/ByteCode/Interp.cpp
@@ -7,7 +7,6 @@
 //===----------------------------------------------------------------------===//
 
 #include "Interp.h"
-#include "ByteCode/Source.h"
 #include "Compiler.h"
 #include "Function.h"
 #include "InterpFrame.h"
@@ -1626,7 +1625,7 @@ bool CheckBitCast(InterpState &S, CodePtr OpPC, const Type *TargetType,
 }
 
 static void compileFunction(InterpState &S, const Function *Func,
-                            SourceLocation Loc) {
+                            CodePtr OpPC) {
 
   const FunctionDecl *Fn = Func->getDecl();
 
@@ -1636,12 +1635,13 @@ static void compileFunction(InterpState &S, const Function *Func,
   // definition to exist or if the existence of the definition affects the
   // semantics of the program.
   if (!Fn->isDefined() && Fn->isImplicitlyInstantiable() && Fn->isConstexpr() &&
-      S.inConstantContext() && !S.TryConstantInitialization &&
+      S.inConstantContext() && !S.PerformingTrialEvaluation &&
       !S.checkingPotentialConstantExpression()) {
     SemaProxy *SP = S.getASTContext().getSemaProxy();
     if (!SP)
       return;
-    SP->InstantiateFunctionDefinition(Loc, const_cast<FunctionDecl *>(Fn));
+    SP->InstantiateFunctionDefinition(S.Current->getLocation(OpPC),
+                                      const_cast<FunctionDecl *>(Fn));
   }
   Fn = Fn->getDefinition();
   if (!Fn)
@@ -1674,7 +1674,7 @@ bool CallVar(InterpState &S, CodePtr OpPC, const Function *Func,
   }
 
   if (!Func->isFullyCompiled())
-    compileFunction(S, Func, S.Current->getLocation(OpPC));
+    compileFunction(S, Func, OpPC);
 
   if (!CheckCallable(S, OpPC, Func))
     return false;
@@ -1751,7 +1751,7 @@ bool Call(InterpState &S, CodePtr OpPC, const Function *Func,
   }
 
   if (!Func->isFullyCompiled())
-    compileFunction(S, Func, S.Current->getLocation(OpPC));
+    compileFunction(S, Func, OpPC);
 
   if (!CheckCallable(S, OpPC, Func))
     return cleanup();