+++ clang/lib/AST/ByteCode/InterpState.cpp
@@ -24,7 +24,7 @@ InterpState::InterpState(const State &Parent, Program &P, InterpStack &Stk,
       StepsLeft(Ctx.getLangOpts().ConstexprStepLimit),
       InfiniteSteps(StepsLeft == 0) {
   InConstantContext = Parent.InConstantContext;
-  TryConstantInitialization = Parent.TryConstantInitialization;
+  PerformingTrialEvaluation = Parent.PerformingTrialEvaluation;
   CheckingPotentialConstantExpression =
       Parent.CheckingPotentialConstantExpression;
   CheckingForUndefinedBehavior = Parent.CheckingForUndefinedBehavior;
@@ -39,7 +39,7 @@ InterpState::InterpState(const State &Parent, Program &P, InterpStack &Stk,
       Current(&BottomFrame), StepsLeft(Ctx.getLangOpts().ConstexprStepLimit),
       InfiniteSteps(StepsLeft == 0) {
   InConstantContext = Parent.InConstantContext;
-  TryConstantInitialization = Parent.TryConstantInitialization;
+  PerformingTrialEvaluation = Parent.PerformingTrialEvaluation;
   CheckingPotentialConstantExpression =
       Parent.CheckingPotentialConstantExpression;
   CheckingForUndefinedBehavior = Parent.CheckingForUndefinedBehavior;